Separated Upload avatar modal from Profile view
